# test-api-server/openapi.yaml
openapi: 3.0.3
info:
  title: MCP Test API Server
  description: |
    A comprehensive API server for testing MCP Studio functionality. 
    Provides CRUD operations for users, chats, messages, and feedbacks with MongoDB backend.

    This API is designed to be easily convertible to an MCP (Model Context Protocol) server.
  version: 1.0.0
  contact:
    name: MCP Studio Team
    email: support@mcpstudio.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.mcpstudio.com
    description: Production server
...
paths:
...
  /api/feedbacks:
    get:
      tags: [feedbacks]
      summary: Get all feedbacks
      description: Retrieve a paginated list of all feedbacks
      operationId: getFeedbacks
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Feedbacks retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
        "500":
          description: Failed to fetch feedbacks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/feedbacks/{id}:
    get:
      tags: [feedbacks]
      summary: Get feedback by ID
      description: Retrieve a specific feedback by its ID
      operationId: getFeedbackById
      parameters:
        - name: id
          in: path
          required: true
          description: Feedback ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Feedback retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        "400":
          description: Invalid feedback ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Feedback not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags: [feedbacks]
      summary: Update feedback
      description: Update an existing feedback
      operationId: updateFeedback
      parameters:
        - name: id
          in: path
          required: true
          description: Feedback ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFeedbackRequest"
      responses:
        "200":
          description: Feedback updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateResponse"
        "400":
          description: Invalid input or feedback ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Feedback not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to update feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [feedbacks]
      summary: Delete feedback
      description: Delete a feedback by its ID
      operationId: deleteFeedback
      parameters:
        - name: id
          in: path
          required: true
          description: Feedback ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Feedback deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "400":
          description: Invalid feedback ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Feedback not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to delete feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/feedbacks/chat/{chatId}:
    get:
      tags: [feedbacks]
      summary: Get feedbacks by chat ID
      description: Retrieve all feedbacks for a specific chat
      operationId: getFeedbacksByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
        - $ref: "#/components/parameters/Page"
        - name: limit
          in: query
          description: Number of feedbacks per page (default 20)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Feedbacks retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackListResponse"
        "500":
          description: Failed to fetch feedbacks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags: [feedbacks]
      summary: Delete all feedbacks for chat
      description: Delete all feedbacks in a specific chat
      operationId: deleteFeedbacksByChatId
      parameters:
        - name: chatId
          in: path
          required: true
          description: Chat ID
          schema:
            type: string
      responses:
        "200":
          description: Feedbacks deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "500":
          description: Failed to delete feedbacks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/feedbacks/user/{userId}:
    get:
      tags: [feedbacks]
      summary: Get user feedback history
      description: Retrieve feedback history for a specific user with filtering and sorting
      operationId: getUserFeedbackHistory
      parameters:
        - name: userId
          in: path
          required: true
          description: User ID (ObjectId)
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: rating
          in: query
          description: Filter by rating
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: isUpvoted
          in: query
          description: Filter by upvote status
          schema:
            type: boolean
        - name: hasJiraIssue
          in: query
          description: Filter by Jira issue presence
          schema:
            type: boolean
        - name: sortBy
          in: query
          description: Sort field
          schema:
            type: string
            enum: [createdAt, rating, lastModifiedAt]
            default: createdAt
        - name: sortOrder
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: User feedback history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserFeedbackHistoryResponse"
        "400":
          description: Invalid user ID format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to fetch user feedback history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    Page:
      name: page
      in: query
      description: Page number for pagination (optional - only used when both page and limit are provided)
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      description: Number of items per page (optional - only used when both page and limit are provided, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
  schemas:
...
    ObjectId:
      type: string
      pattern: "^[0-9a-fA-F]{24}$"
      description: MongoDB ObjectId
    Timestamp:
      type: string
      format: date-time
      description: ISO 8601 timestamp

    # Error responses
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Detailed error message

    # Pagination (optional - only when page and limit are both provided)
    PaginationOptional:
      type: object
      properties:
        current:
          type: integer
          description: Current page number
        total:
          type: integer
          description: Total number of pages
        count:
          type: integer
          description: Number of items in current page
        totalRecords:
          type: integer
          description: Total number of records

    # Base list response without pagination
...
    JiraIssue:
      type: object
      required:
        - issueKey
        - issueId
        - issueUrl
        - summary
        - description
        - status
        - priority
        - labels
        - createdDate
      properties:
        issueKey:
          type: string
        issueId:
          type: string
        issueUrl:
          type: string
        summary:
          type: string
        description:
          type: string
        status:
          type: string
        assignee:
          type: string
        priority:
          type: string
        labels:
          type: array
          items:
            type: string
        attachments:
          type: array
          items:
            type: string
        createdDate:
          $ref: "#/components/schemas/Timestamp"
    Feedback:
      type: object
      required:
        - _id
        - chatId
        - messageId
        - userId
        - comments
        - rating
        - isUpvoted
        - hasJiraIssue
        - createdAt
        - lastModifiedAt
      properties:
        _id:
          $ref: "#/components/schemas/ObjectId"
        chatId:
          type: string
        messageId:
          type: string
        userId:
          $ref: "#/components/schemas/ObjectId"
        comments:
          type: string
        rating:
          type: integer
          minimum: 0
          maximum: 5
        isUpvoted:
          type: boolean
        hasJiraIssue:
          type: boolean
        jiraIssue:
          $ref: "#/components/schemas/JiraIssue"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
        lastModifiedAt:
          $ref: "#/components/schemas/Timestamp"
...
    UpdateFeedbackRequest:
      type: object
      properties:
        comments:
          type: string
        rating:
          type: integer
          minimum: 0
          maximum: 5
        isUpvoted:
          type: boolean
        hasJiraIssue:
          type: boolean
        jiraIssue:
          $ref: "#/components/schemas/JiraIssue"
...
    FeedbackListResponse:
      type: object
      required:
        - feedbacks
        - totalRecords
      properties:
        feedbacks:
          type: array
          items:
            $ref: "#/components/schemas/Feedback"
        totalRecords:
          type: integer
          description: Total number of feedbacks in the collection
        pagination:
          $ref: "#/components/schemas/PaginationOptional"
    UserFeedbackHistoryResponse:
      type: object
      required:
        - feedbacks
        - totalRecords
      properties:
        feedbacks:
          type: array
          items:
            allOf:
              - $ref: "#/components/schemas/Feedback"
              - type: object
                properties:
                  id:
                    type: string
        totalRecords:
          type: integer
          description: Total number of feedbacks for the user
        pagination:
          allOf:
            - $ref: "#/components/schemas/PaginationOptional"
            - type: object
              properties:
                hasNextPage:
                  type: boolean
                hasPreviousPage:
                  type: boolean

    # Generic response schemas
    UpdateResponse:
      type: object
      required:
        - message
        - modifiedCount
      properties:
        message:
          type: string
        modifiedCount:
          type: integer
    DeleteResponse:
      type: object
      required:
        - message
        - deletedCount
      properties:
        message:
          type: string
        deletedCount:
          type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - bearerAuth: []